Conversión a BigDecimal

+++ demo/src/main/java/com/example/demo/application/ingredientApplication/IngredientDTO.java
@@ -1,11 +1,12 @@
 package com.example.demo.application.ingredientApplication;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class IngredientDTO implements Serializable{
     
     public UUID id;
     public String name;
-    public double price;
+    public BigDecimal price;
 }
